+++ spec/truffle/array/append_spec.rb
@@ -20,7 +20,7 @@ describe "Array#<<" do
 
   it "empty array has null storage" do
     ary = []
-    storage(ary).should == "null"
+    storage(ary).should == "empty"
   end
 
   it "supports transitions from null storage" do